Absolluting arousal response
Given the that under some views the phrasing of the arousal question can be seen as dichotomic, we can absolute it if don't care about the direction, while preserving the intensity. This seems to work well, but need to see if it changes teh relationship with other variables of interest (e.g. correlation between subjective and objective arousal