bootstrap for PCA/MFA
began the fucntions to do boostrap tests to ases stability of PCA/MFA results.

we could do a regular test like the probaility of more extreme values (eigen bectors.values) to the observed ones, but not sure this is appropriate.

A common option in papers is to compute confidence intervals around the estimate, to check if the values are above a certain treshold, e,g  5% 95% for eigenvalue of 1